recognition/requirements.txt
tensorflow
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
keras
flask
pickle-mixin